OOP patch for a typelevel project ¯\_(ツ)_/¯
