+++ jetty/src/test/scala/org/http4s/server/jetty/Issue454.scala
@@ -7,7 +7,7 @@ import cats.effect.IO
 import org.eclipse.jetty.server.{HttpConfiguration, HttpConnectionFactory, Server, ServerConnector}
 import org.eclipse.jetty.servlet.{ServletContextHandler, ServletHolder}
 import org.http4s.dsl.io._
-import org.http4s.servlet.Http4sServlet
+import org.http4s.servlet.AsyncHttp4sServlet
 
 object Issue454 {
   // If the bug is not triggered right away, try increasing or
@@ -41,7 +41,7 @@ object Issue454 {
     server.start()
   }
 
-  val servlet = new Http4sServlet[IO](
+  val servlet = new AsyncHttp4sServlet[IO](
     service = HttpRoutes.of[IO] {
       case GET -> Root => Ok(insanelyHugeData)
     },
