+++ servlet/src/main/scala/org/http4s/servlet/Http4sServlet.scala
@@ -1,36 +1,28 @@
-package org.http4s
-package servlet
+package org.http4s.servlet
 
-import cats.data.OptionT
-import cats.effect._
-import cats.implicits.{catsSyntaxEither => _, _}
-import fs2.async
 import java.net.InetSocketAddress
-import javax.servlet._
+import javax.servlet.ServletConfig
 import javax.servlet.http.{HttpServlet, HttpServletRequest, HttpServletResponse, HttpSession}
+
+import cats.effect.Effect
+import cats.implicits._
+import org.http4s._
 import org.http4s.headers.`Transfer-Encoding`
-import org.http4s.server._
+import org.http4s.server.ServerSoftware
 import org.log4s.getLogger
+
 import scala.collection.JavaConverters._
-import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.Duration
 
-class Http4sServlet[F[_]](
-    service: HttpRoutes[F],
-    asyncTimeout: Duration = Duration.Inf,
-    implicit private[this] val executionContext: ExecutionContext = ExecutionContext.global,
-    private[this] var servletIo: ServletIo[F],
-    serviceErrorHandler: ServiceErrorHandler[F])(implicit F: Effect[F])
+abstract class Http4sServlet[F[_]](service: HttpRoutes[F], servletIo: ServletIo[F])(
+    implicit F: Effect[F])
     extends HttpServlet {
-  private[this] val logger = getLogger
-
-  private val asyncTimeoutMillis = if (asyncTimeout.isFinite()) asyncTimeout.toMillis else -1 // -1 == Inf
-
-  private[this] var serverSoftware: ServerSoftware = _
+  protected val logger = getLogger
 
   // micro-optimization: unwrap the service and call its .run directly
-  private[this] val serviceFn = service.run
-  private[this] val optionTSync = Sync[OptionT[F, ?]]
+  protected val serviceFn = service.run
+
+  protected var servletApiVersion: ServletApiVersion = _
+  private[this] var serverSoftware: ServerSoftware = _
 
   object ServletRequestKeys {
     val HttpSession: AttributeKey[Option[HttpSession]] = AttributeKey[Option[HttpSession]]
@@ -38,52 +30,12 @@ class Http4sServlet[F[_]](
 
   override def init(config: ServletConfig): Unit = {
     val servletContext = config.getServletContext
-    val servletApiVersion = ServletApiVersion(servletContext)
+    servletApiVersion = ServletApiVersion(servletContext)
     logger.info(s"Detected Servlet API version $servletApiVersion")
-
-    verifyServletIo(servletApiVersion)
-    logServletIo()
     serverSoftware = ServerSoftware(servletContext.getServerInfo)
   }
 
-  // TODO This is a dodgy check.  It will have already triggered class loading of javax.servlet.WriteListener.
-  // Remove when we can break binary compatibility.
-  private def verifyServletIo(servletApiVersion: ServletApiVersion): Unit = servletIo match {
-    case NonBlockingServletIo(chunkSize) if servletApiVersion < ServletApiVersion(3, 1) =>
-      logger.warn(
-        "Non-blocking servlet I/O requires Servlet API >= 3.1. Falling back to blocking I/O.")
-      servletIo = BlockingServletIo[F](chunkSize)
-    case _ => // cool
-  }
-
-  private def logServletIo(): Unit =
-    logger.info(servletIo match {
-      case BlockingServletIo(chunkSize) => s"Using blocking servlet I/O with chunk size $chunkSize"
-      case NonBlockingServletIo(chunkSize) =>
-        s"Using non-blocking servlet I/O with chunk size $chunkSize"
-    })
-
-  override def service(
-      servletRequest: HttpServletRequest,
-      servletResponse: HttpServletResponse): Unit =
-    try {
-      val ctx = servletRequest.startAsync()
-      ctx.setTimeout(asyncTimeoutMillis)
-      // Must be done on the container thread for Tomcat's sake when using async I/O.
-      val bodyWriter = servletIo.initWriter(servletResponse)
-      async.unsafeRunAsync(
-        toRequest(servletRequest).fold(
-          onParseFailure(_, servletResponse, bodyWriter),
-          handleRequest(ctx, _, bodyWriter)
-        )) {
-        case Right(()) =>
-          IO(ctx.complete())
-        case Left(t) =>
-          IO(errorHandler(servletRequest, servletResponse)(t))
-      }
-    } catch errorHandler(servletRequest, servletResponse)
-
-  private def onParseFailure(
+  protected def onParseFailure(
       parseFailure: ParseFailure,
       servletResponse: HttpServletResponse,
       bodyWriter: BodyWriter[F]): F[Unit] = {
@@ -94,47 +46,7 @@ class Http4sServlet[F[_]](
     renderResponse(response, servletResponse, bodyWriter)
   }
 
-  private def handleRequest(
-      ctx: AsyncContext,
-      request: Request[F],
-      bodyWriter: BodyWriter[F]): F[Unit] = {
-    ctx.addListener(new AsyncTimeoutHandler(request, bodyWriter))
-    // Note: We're catching silly user errors in the lift => flatten.
-    val response = Async.shift(executionContext) *>
-      optionTSync
-        .suspend(serviceFn(request))
-        .getOrElse(Response.notFound)
-        .recoverWith(serviceErrorHandler(request))
-
-    val servletResponse = ctx.getResponse.asInstanceOf[HttpServletResponse]
-    renderResponse(response, servletResponse, bodyWriter)
-  }
-
-  private class AsyncTimeoutHandler(request: Request[F], bodyWriter: BodyWriter[F])
-      extends AbstractAsyncListener {
-    override def onTimeout(event: AsyncEvent): Unit = {
-      val ctx = event.getAsyncContext
-      val servletResponse = ctx.getResponse.asInstanceOf[HttpServletResponse]
-      async.unsafeRunAsync {
-        if (!servletResponse.isCommitted) {
-          val response =
-            F.pure(
-              Response[F](Status.InternalServerError)
-                .withEntity("Service timed out."))
-          renderResponse(response, servletResponse, bodyWriter)
-        } else {
-          logger.warn(
-            s"Async context timed out, but response was already committed: ${request.method} ${request.uri.path}")
-          F.unit
-        }
-      } {
-        case Right(()) => IO(ctx.complete())
-        case Left(t) => IO(logger.error(t)("Error timing out async context")) *> IO(ctx.complete())
-      }
-    }
-  }
-
-  private def renderResponse(
+  protected def renderResponse(
       response: F[Response[F]],
       servletResponse: HttpServletResponse,
       bodyWriter: BodyWriter[F]): F[Unit] =
@@ -156,27 +68,7 @@ class Http4sServlet[F[_]](
         }
     }
 
-  private def errorHandler(
-      servletRequest: ServletRequest,
-      servletResponse: HttpServletResponse): PartialFunction[Throwable, Unit] = {
-    case t: Throwable if servletResponse.isCommitted =>
-      logger.error(t)("Error processing request after response was committed")
-
-    case t: Throwable =>
-      logger.error(t)("Error processing request")
-      val response = F.pure(Response[F](Status.InternalServerError))
-      // We don't know what I/O mode we're in here, and we're not rendering a body
-      // anyway, so we use a NullBodyWriter.
-      async
-        .unsafeRunAsync(renderResponse(response, servletResponse, NullBodyWriter)) { _ =>
-          IO {
-            if (servletRequest.isAsyncStarted)
-              servletRequest.getAsyncContext.complete()
-          }
-        }
-  }
-
-  private def toRequest(req: HttpServletRequest): ParseResult[Request[F]] =
+  protected def toRequest(req: HttpServletRequest): ParseResult[Request[F]] =
     for {
       method <- Method.fromString(req.getMethod)
       uri <- Uri.requestTarget(
@@ -206,25 +98,11 @@ class Http4sServlet[F[_]](
         )
       )
 
-  private def toHeaders(req: HttpServletRequest): Headers = {
+  protected def toHeaders(req: HttpServletRequest): Headers = {
     val headers = for {
       name <- req.getHeaderNames.asScala
       value <- req.getHeaders(name).asScala
     } yield Header(name, value)
     Headers(headers.toSeq: _*)
   }
-}
-
-object Http4sServlet {
-  def apply[F[_]: Effect](
-      service: HttpRoutes[F],
-      asyncTimeout: Duration = Duration.Inf,
-      executionContext: ExecutionContext = ExecutionContext.global): Http4sServlet[F] =
-    new Http4sServlet[F](
-      service,
-      asyncTimeout,
-      executionContext,
-      BlockingServletIo[F](DefaultChunkSize),
-      DefaultServiceErrorHandler
-    )
 }
